create measure table component
