+++ pages/measures/[id].jsx
@@ -1,15 +1,11 @@
 import { useRouter } from 'next/router';
 import useSWR from 'swr';
 import Link from 'next/link';
-import Image from 'next/image';
 import Layout from '../../components/layouts/layout';
 import LinkIcon from '../../components/link-icon';
 import MeasureOverview from '../../components/measure-overview';
+import MeasureTable from '../../components/measure-table';
 
-const formatDate = (date) => {
-  let dateObject = new Date(date);
-  return dateObject.toLocaleDateString();
-};
 
 const checkURL = (text) => {
   let match = text.match(
@@ -136,69 +132,7 @@ export default function Law() {
                 <h2 className='pb-2 mobile sm:main'>{data.kop_5_juridische_toelichting}</h2>
                 <p className='body-text-mobile sm:body-text'>{data.juridische_toelichting}</p>
               </div>
-
-              {/* refactor table to have different design on mobile */}
-              <table className='table-fixed w-full mt-5'>
-                <tbody>
-                  <tr className='my-10 border-b-2 border-t-2'>
-                    <td className='w-1/2 font-manrope text-base font-normal'>Rechtsgebied</td>
-                    <td className='w-1/2 font-manrope text-base font-bold'>
-                      {data.rechtsgebied} - {data.subrechtsgebied}
-                    </td>
-                  </tr>
-                  <tr className='my-10 border-b-2'>
-                    <td className='w-1/2 font-manrope text-base font-normal'>Citeertitel</td>
-                    <td className='w-1/2 font-manrope text-base font-bold'>{data.citeertitel}</td>
-                  </tr>
-                  <tr className='my-10 border-b-2'>
-                    <td className='w-1/2 font-manrope text-base font-normal'>Artikel</td>
-                    <td className='w-1/2 font-manrope text-base font-bold'>
-                      <a
-                        className='text-greenLink underline'
-                        target='_blank'
-                        href={data.link_naar_wetsartikel}
-                        rel='noreferrer'
-                      >
-                        {data.artikel}
-                        <span className='pl-2'>
-                          <Image
-                            alt='new tab'
-                            src='/icons/Vectorlink-icon.svg'
-                            width={13}
-                            height={13}
-                          />
-                        </span>
-                      </a>
-                    </td>
-                  </tr>
-                  <tr className='my-10 border-b-2'>
-                    <td className='w-1/2 font-manrope text-base font-normal'>Geldig vanaf</td>
-                    <td className='w-1/2 font-manrope text-base font-bold'>
-                      {formatDate(data.geldend_vanaf)}
-                    </td>
-                  </tr>
-                  <tr className='my-10 border-b-2'>
-                    <td className='w-1/2 font-manrope text-base font-normal'>Bevoegdheidsniveau</td>
-                    <td className='w-1/2 font-manrope text-base font-bold'>
-                      {data.europees && <span>Europees </span>}
-                      {data.nationaal && <span>Nationaal </span>}
-                      {data.provinciaal && <span>Provinciaal </span>}
-                      {data.waterschappen && <span>Waterschappen </span>}
-                      {data.gemeentelijk && <span>Gemeentelijk</span>}
-                    </td>
-                  </tr>
-                  <tr className='my-10 border-b-2'>
-                    <td className='w-1/2 font-manrope text-base font-normal'>Type document</td>
-                    <td className='w-1/2 font-manrope text-base font-bold'>{data.type_document}</td>
-                  </tr>
-                  <tr>
-                    <td className='font-manrope text-base font-normal'>Beleidsinstrument</td>
-                    <td className='font-manrope text-base font-bold'>
-                      {data.type_beleidsinstrument}
-                    </td>
-                  </tr>
-                </tbody>
-              </table>
+                <MeasureTable data={data} />
             </div>
 
             <MeasureOverview data={data} viewport='mobile' />
